+++ .changes/types-package.md
@@ -1,0 +1,11 @@
+---
+"covector": minor
+"@covector/apply": minor
+"@covector/assemble": minor
+"@covector/changelog": minor
+"@covector/command": minor
+"@covector/files": minor
+"@covector/types": patch
+---
+
+Extract out types into separate package to remove issues with circular dependencies.
