+++ package.json
@@ -28,6 +28,6 @@
   },
   "volta": {
     "node": "14.17.0",
-    "npm": "7.15.0"
+    "npm": "7.18.1"
   }
 }
