+++ packages/action/package.json
@@ -17,6 +17,7 @@
     "@effection/node": "1.0.2"
   },
   "devDependencies": {
+    "@covector/types": "^0.0.0",
     "@vercel/ncc": "^0.27.0",
     "fixturez": "^1.1.0",
     "tslib": "^2.2.0",
