+++ packages/changelog/package.json
@@ -24,6 +24,7 @@
     "unified": "^9.2.1"
   },
   "devDependencies": {
+    "@covector/types": "^0.0.0",
     "@types/lodash": "^4.14.168",
     "tslib": "^2.2.0",
     "typescript": "^4.2.3"
