+++ packages/changelog/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.base.json",
-  "references": [{ "path": "../files" }],
+  "references": [{ "path": "../files" }, { "path": "../types" }],
   "compilerOptions": {
     "baseUrl": ".",
     "outDir": "dist",
