+++ packages/command/src/index.ts
@@ -4,20 +4,12 @@ import execa from "execa";
 import stripAnsi from "strip-ansi";
 import path from "path";
 
-import type { PkgVersion, PkgPublish } from "@covector/assemble";
-
-type RunningCommand = {
-  command?: string | Function;
-  shouldRunCommand?: boolean;
-  runFromRoot?: boolean;
-};
-
-type NormalizedCommand = {
-  command?: string;
-  runFromRoot?: boolean;
-  dryRunCommand?: boolean;
-  pipe?: boolean;
-};
+import type {
+  PkgVersion,
+  PkgPublish,
+  RunningCommand,
+  NormalizedCommand,
+} from "@covector/types";
 
 export const attemptCommands = function* ({
   cwd,
