+++ packages/covector/package.json
@@ -33,6 +33,7 @@
     "yargs": "^15.3.1"
   },
   "devDependencies": {
+    "@covector/types": "^0.0.0",
     "@types/inquirer": "^7.3.1",
     "fixturez": "^1.1.0",
     "tslib": "^2.2.0",
