+++ packages/files/tsconfig.json
@@ -5,5 +5,6 @@
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts"],
+  "references": [{ "path": "../types" }]
 }
