+++ packages/types/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@covector/apply",
-  "version": "0.4.0",
+  "name": "@covector/types",
+  "version": "0.0.0",
   "license": "Apache-2.0",
   "homepage": "https://github.com/jbolda/covector#readme",
   "author": "Jacob Bolda <me@jacobbolda.com> (https://www.jacobbolda.com/)",
@@ -9,20 +9,12 @@
     "url": "https://github.com/jbolda/covector.git"
   },
   "main": "dist/index.js",
-  "types": "dist/index.d.ts",
   "scripts": {
-    "build": "tsc -b",
-    "clean": "rimraf dist node_modules",
-    "prepublishOnly": "tsc -b"
-  },
-  "dependencies": {
-    "@covector/files": "0.4.0",
-    "lodash": "^4.17.21",
-    "semver": "^7.3.2"
+    "build": "echo noop",
+    "clean": "echo noop"
   },
   "devDependencies": {
     "@covector/types": "^0.0.0",
-    "@types/lodash": "^4.14.168",
     "tslib": "^2.2.0",
     "typescript": "^4.2.3"
   },
