+++ tsconfig.base.json
@@ -14,12 +14,13 @@
     "declaration": true,
     "declarationMap": true,
     "paths": {
+      "@covector/types": ["packages/types/src"],
       "covector": ["packages/covector/src"],
-      "@covector/apply": ["packages/apply/src"],
-      "@covector/assemble": ["packages/assemble/src"],
-      "@covector/changelog": ["packages/changelog/src"],
+      "@covector/files": ["packages/files/src"],
       "@covector/command": ["packages/command/src"],
-      "@covector/files": ["packages/files/src"]
+      "@covector/changelog": ["packages/changelog/src"],
+      "@covector/assemble": ["packages/assemble/src"],
+      "@covector/apply": ["packages/apply/src"]
     }
   }
 }
